requirements.txt
Django==4.1.3
python-decouple==3.6
psycopg2-binary==2.9.5
djangorestframework==3.14.0
drf-yasg==1.21.7